pyproject.toml
[build-system]
requires = ["hatchling>=1.27"]
build-backend = "hatchling.build"

[project]
name = "quant-os"
version = "0.1.0"
description = "Quant operating system MVP foundation"
readme = "README.md"
requires-python = ">=3.12,<3.15"
license = { text = "MIT" }
authors = [
  { name = "Lia" },
]
dependencies = [
  "alembic>=1.16,<2",
  "duckdb>=1.4,<2",
  "fastapi>=0.116,<1",
  "pydantic>=2.11,<3",
  "PyYAML>=6.0,<7",
  "SQLAlchemy>=2.0,<3",
  "typer>=0.16,<1",
  "uvicorn>=0.35,<1",
]

[project.optional-dependencies]
dev = [
  "httpx>=0.28,<1",
  "pytest>=8.4,<9",
]

[project.scripts]
quant-os = "quant_os.cli.main:main"

[tool.hatch.build.targets.wheel]
packages = ["src/quant_os"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
